Rebase on Centos8, use Spectrum Scale 5.1.1.0 Developer Edition and add CES/Object and KVM/Libvirt support.
This major update introduces support for vagrant-libvirt so that one can use the native Linux KVM virtualization
instead of Virtualbox on Linux machines, like tested on Fedora 33.

Support for Spectrum Scale Developer Edition is also added as it is available as free download. With that we are
also introducing the 5.1.1.0 version that comes with the new Ansible-based installer.

In order to be able to explore the new Spectrum Scale Object stack as introduced with Spectrum Scale 5.1.0.1, we
needed to re-base the project on Centos 8.

Finally the requirements for RAM have been increased, the disk and filesystem configuration has been changed and
an additional network got added so that we can enable Cluster Export Services and untimately install Spectrum Scale
Object as additional functionality.

The included Spectrum Scale Object functionality includes support for the S3 protocol. A brief configuration demo
showing how to set up S3 credentials is also added.