Fix #32, adding /usr/local/bin to the path to support pip-based Ansible installation.
This change is addressing https://github.com/IBM/SpectrumScaleVagrant/issues/32
Due to the changed prep-box with ansible no longer installed as RPM, but using
pip, the executables are located in /usr/local/bin. That directory needs to
be added to the path when running spectrumscale as sudo during deployment so
that the ansible-based spectrumscale installer can call ansible-playbook.